+++ backend/src/main/java/com/example/RecruitEmployee/project/ProjectService.java
@@ -1,9 +1,10 @@
 package com.example.RecruitEmployee.project;
 
-import com.example.RecruitEmployee.dto.EmployeeDto;
 import com.example.RecruitEmployee.dto.ProjectDto;
 import com.example.RecruitEmployee.employee.Employee;
 import com.example.RecruitEmployee.exception.ApiRequestException;
+import com.example.RecruitEmployee.mapper.EmployeeMapper;
+import com.example.RecruitEmployee.mapper.ManagerMapper;
 import com.example.RecruitEmployee.mapper.ProjectMapper;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
@@ -20,11 +21,15 @@ import java.util.stream.Collectors;
 public class ProjectService {
     private final ProjectMapper projectMapper;
     private final ModelMapper modelMapper;
+    private final ManagerMapper managerMapper;
+    private final EmployeeMapper employeeMapper;
 
     @Autowired
-    public ProjectService(@Qualifier("ProjectMapper") ProjectMapper projectMapper, ModelMapper modelMapper) {
+    public ProjectService(@Qualifier("ProjectMapper") ProjectMapper projectMapper, ModelMapper modelMapper, ManagerMapper managerMapper, EmployeeMapper employeeMapper) {
         this.projectMapper = projectMapper;
         this.modelMapper = modelMapper;
+        this.managerMapper = managerMapper;
+        this.employeeMapper = employeeMapper;
     }
 
     public List<ProjectDto> getAllProject() {
@@ -50,24 +55,24 @@ public class ProjectService {
         return convertToDto(projectMapper.getProjectById(projectId).orElseThrow(() -> new ApiRequestException("Project not found with projectId:" + projectId)));
     }
 
-    public int addProject(ProjectDto project) {
+    public int addProject(Project project) {
         Optional<Project> project1 = projectMapper.getProjectByName(project.getProjectName());
+        managerMapper.getManagerById(project.getManagerId()).orElseThrow(()->new ApiRequestException("No Manager with Manager Id:" + project.getManagerId()));
         if(project1.isPresent()) throw new ApiRequestException("Project already exist");
-        return projectMapper.addProject(covertToEntity(project));
+        return projectMapper.addProject(project);
     }
 
-    public int updateProject(ProjectDto project) {
+    public int updateProject(Project project) {
         projectMapper.getProjectById(project.getProjectId()).orElseThrow(() -> new ApiRequestException("Project not found with projectId:" + project.getProjectId()));
-        return projectMapper.updateProject(covertToEntity(project));
+        managerMapper.getManagerById(project.getManagerId()).orElseThrow(()->new ApiRequestException("No Manager with Manager Id:" + project.getManagerId()));
+        return projectMapper.updateProject(project);
     }
 
     public int deleteProject(Integer projectId) {
         projectMapper.getProjectById(projectId).orElseThrow(() -> new ApiRequestException("Project not found with projectId:" + projectId));
-        try {
-            return projectMapper.deleteProject(projectId);
-        } catch (ApiRequestException e){
-            throw new ApiRequestException("Cannot delete Project, since it's linked to Employees");
-        }
+        List<Employee> employees =employeeMapper.getEmployeeByProjectId(projectId);
+        if(!employees.isEmpty()) throw new ApiRequestException("Cannot delete Project, since it's linked to Employees");
+        return projectMapper.deleteProject(projectId);
     }
 
     public List<ProjectDto> getProjectByManagerId(Integer managerId) {
@@ -83,9 +88,4 @@ public class ProjectService {
     private ProjectDto convertToDto(Project project){
         return modelMapper.map(project, ProjectDto.class);
     }
-
-    private Project covertToEntity(ProjectDto projectDto){
-        return modelMapper.map(projectDto, Project.class);
-    }
-
 }
